# Move automtic simplifications of gamma into a calculate function