--- a/symengine/tests/basic/CMakeLists.txt
+++ b/symengine/tests/basic/CMakeLists.txt
@@ -125,6 +125,10 @@ add_executable(test_assumptions test_assumptions.cpp)
 target_link_libraries(test_assumptions symengine catch)
 add_test(test_assumptions ${PROJECT_BINARY_DIR}/test_assumptions)
 
+add_executable(test_calculate test_calculate.cpp)
+target_link_libraries(test_calculate symengine catch)
+add_test(test_calculate ${PROJECT_BINARY_DIR}/test_calculate)
+
 add_executable(test_refine test_refine.cpp)
 target_link_libraries(test_refine symengine catch)
 add_test(test_refine ${PROJECT_BINARY_DIR}/test_refine)